more usage of unified entrypoint.sh

+++ roles/container/templates/entrypoint.sh
@@ -3,6 +3,9 @@
 exitcode=1
 watched_pids=""
 
+{% if entrypoint_pre is defined %}
+{{entrypoint_pre}}
+{% endif %}
 ignore()
 {
     echo "Entrypoint - Ignoring SIGHUP"
@@ -16,12 +19,18 @@ stop()
 
     exitcode=0
 
-    echo "Entrypoint - Send 'TERM' to pid(s) '$watched_pids'"
-    kill -s TERM $watched_pids
+{% if entrypoint_shutdown is defined %}{% for cmdline in entrypoint_shutdown %}
+    {{ cmdline }}
+{% endfor %}{% endif %}
 
-    # No need to wait. Otherwise "Terminated" log message will occur in journald
-    #echo "Entrypoint - Wait for pid(s) '$watched_pids'"
-    #wait $watched_pids
+    if [ ! -z "$watched_pids" ]; then
+        echo "Entrypoint - Send 'TERM' to pid(s) '$watched_pids'"
+        kill -s TERM $watched_pids
+
+        # No need to wait. Otherwise "Terminated" log message will occur in journald
+        #echo "Entrypoint - Wait for pid(s) '$watched_pids'"
+        #wait $watched_pids
+    fi
 
     #echo "Entrypoint - Exit $exitcode"
     #exit $exitcode
@@ -32,7 +41,7 @@ trap "stop" SIGTERM SIGINT
 start()
 {
     echo "Entrypoint - Starting service"
-{% for cmdline in entrypoint_services %}
+{% for cmdline in entrypoint_startup %}
     {{ cmdline }}
 {% if cmdline[-2:] == ' &' %}
     PID=$!
@@ -40,13 +49,19 @@ start()
 {% endif %}
 {% endfor %}
 
-    watched_pids=$(echo $watched_pids | xargs)
-
-    echo "Entrypoint - Service started with pid(s) '$watched_pids'"
+    if [ ! -z "$watched_pids" ]; then
+        watched_pids=$(echo $watched_pids | xargs)
+        echo "Entrypoint - Service started with pid(s) '$watched_pids'"
+    else
+        echo "Entrypoint - Service started"
+    fi
 }
 
 start
 
+{% if entrypoint_check is defined %}
+    {{entrypoint_check}}
+{% else %}
 if [ ! -z "$watched_pids" ]; then
     echo "Entrypoint - Observe pid(s) '$watched_pids'"
 
@@ -64,6 +79,7 @@ if [ ! -z "$watched_pids" ]; then
     while eval "$CMD"; do sleep 1 & wait; done
     #while true; do wait $watched_pids; break; done
 fi
+{% endif %}
 
 if [ $exitcode -ne 0 ]; then
     echo "Entrypoint - Unexpected interruption with code '$exitcode'"
